backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
supabase
python-dotenv
torch
torchvision
google-generativeai
